fix: move slack oauth endpoint to server and other imrpovements
